+++ lib/fixtures.go
@@ -63,6 +63,7 @@ func DeployConfigFixture(name string) DeployConfig {
 			Startup: Startup{
 				SkipCheck: true,
 			},
+			SingularityRequestID: "blah",
 		}
 	}
 }